Fix coin package dao inproper data type of cost.

+++ internal/app/contracts/coin.go
@@ -9,7 +9,7 @@ type OrderCoinParams struct {
 	BuyerID     int
 	PackageId   int
 	Quantity    int
-	Cost        int
+	Cost        float64
 	OrderStatus models.OrderStatus
 }
